[#PRX-250 Possible race condition for receiving n-event-rsp]
Solve by copying naction files before actually receiving a valid
response from the destination in both direct and scheduled forwarding.
In case the reponse is success then the naction dir copy is removed
while keeping the copy in the nevent dir, in case of callingAET of the
naction not equal to that of the proxy then the proxy should not receive
nevent reports and therefore the copy in the naction and nevent dirs are
removed, in case of failure the naction file is renamed according to
error and the copy from the nevent dir is removed.